Rename AuthProvider, AuthContext and useAuth to Auth0Provider, Auth0Context and useAuth0
