+++ __tests__/use-auth.test.tsx
@@ -1,14 +1,14 @@
-import useAuth from '../src/use-auth';
+import useAuth0 from '../src/use-auth0';
 import { renderHook } from '@testing-library/react-hooks';
 import { createWrapper } from './helpers';
 
-describe('useAuth', () => {
+describe('useAuth0', () => {
   it('should provide the auth context', async () => {
     const wrapper = createWrapper();
     const {
       result: { current },
       waitForNextUpdate,
-    } = renderHook(useAuth, { wrapper });
+    } = renderHook(useAuth0, { wrapper });
     await waitForNextUpdate();
     expect(current).toBeDefined();
   });
@@ -16,9 +16,9 @@ describe('useAuth', () => {
   it('should throw with no provider', () => {
     const {
       result: { current },
-    } = renderHook(useAuth);
+    } = renderHook(useAuth0);
     expect(current.login).toThrowError(
-      'You forgot to wrap your component in <AuthProvider>.'
+      'You forgot to wrap your component in <Auth0Provider>.'
     );
   });
 });
