+++ src/auth0-context.tsx
@@ -2,7 +2,7 @@ import { LogoutOptions, RedirectLoginOptions } from '@auth0/auth0-spa-js';
 import { createContext } from 'react';
 import { AuthState, initialAuthState } from './auth-state';
 
-export interface AuthContextInterface extends AuthState {
+export interface Auth0ContextInterface extends AuthState {
   /**
    * Login in with a redirect.
    */
@@ -15,10 +15,10 @@ export interface AuthContextInterface extends AuthState {
 }
 
 const stub = (): never => {
-  throw new Error('You forgot to wrap your component in <AuthProvider>.');
+  throw new Error('You forgot to wrap your component in <Auth0Provider>.');
 };
 
-export default createContext<AuthContextInterface>({
+export default createContext<Auth0ContextInterface>({
   ...initialAuthState,
   login: stub,
   logout: stub,
