--- src/use-auth.tsx
@@ -1,4 +1,0 @@
-import { useContext } from 'react';
-import AuthContext, { AuthContextInterface } from './auth-context';
-
-export default (): AuthContextInterface => useContext(AuthContext);
